# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sample CRUD api for estuary db",
        "title": "Sample CRUD api for estuary db",
        "contact": {
            "name": "Me",
            "url": "http://me.com/terms.html",
            "email": "alvin@protocol.ai"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/contentwalletlogs": {
            "get": {
                "description": "GetAllContentWalletLogs is a handler to get a slice of record(s) from content_wallet_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentWalletLogs"
                ],
                "summary": "Get list of ContentWalletLogs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page requested (defaults to 0)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "number of records in a page  (defaults to 20)",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "db sort order column",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.PagedResults"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ContentWalletLogs"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "add to add a single record to content_wallet_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentWalletLogs"
                ],
                "summary": "Add an record to content_wallet_logs table",
                "parameters": [
                    {
                        "description": "Add ContentWalletLogs",
                        "name": "ContentWalletLogs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ContentWalletLogs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContentWalletLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/contentwalletlogs/{contenWalletLogsID}": {
            "get": {
                "description": "GetContentWalletLogs is a function to get a single record from the content_wallet_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentWalletLogs"
                ],
                "summary": "Get record from table ContentWalletLogs by  contenWalletLogsID",
                "operationId": "contenWalletLogsID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "contenWalletLogsID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContentWalletLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "ErrNotFound, db record for id not found - returns NotFound HTTP 404 not found error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a single record from content_wallet_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentWalletLogs"
                ],
                "summary": "Update an record in table content_wallet_logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "contenWalletLogsID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update ContentWalletLogs record",
                        "name": "ContentWalletLogs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ContentWalletLogs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContentWalletLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a single record from content_wallet_logs table in the estuary database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentWalletLogs"
                ],
                "summary": "Delete a record from content_wallet_logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "contenWalletLogsID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/model.ContentWalletLogs"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "api.PagedResults": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalRecords": {
                    "type": "integer"
                }
            }
        },
...
        "model.ContentWalletLogs": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "[ 1] content                                        INT8                 null: true   primary: false  isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "$ref": "#/definitions/null.Int"
                },
                "createdAt": {
                    "description": "[ 7] created_at                                     TIMESTAMPTZ          null: true   primary: false  isArray: false  auto: false  col: TIMESTAMPTZ     len: -1      default: []",
                    "type": "string"
                },
                "deltaNodeUUID": {
                    "description": "[ 9] delta_node_uuid                                TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "id": {
                    "description": "[ 0] id                                             INT8                 null: false  primary: true   isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "type": "integer"
                },
                "nodeInfo": {
                    "description": "[ 3] node_info                                      TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "requesterInfo": {
                    "description": "[ 4] requester_info                                 TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "requestingAPIKey": {
                    "description": "[ 5] requesting_api_key                             TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "systemContentWalletID": {
                    "description": "[ 6] system_content_wallet_id                       INT8                 null: true   primary: false  isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "$ref": "#/definitions/null.Int"
                },
                "updatedAt": {
                    "description": "[ 8] updated_at                                     TIMESTAMPTZ          null: true   primary: false  isArray: false  auto: false  col: TIMESTAMPTZ     len: -1      default: []",
                    "type": "string"
                },
                "wallet": {
                    "description": "[ 2] wallet                                         TEXT                 null: true   primary: false  isArray: false  auto: false  col: TEXT            len: -1      default: []",
                    "$ref": "#/definitions/null.String"
                },
                "walletID": {
                    "description": "[10] wallet_id                                      INT8                 null: true   primary: false  isArray: false  auto: false  col: INT8            len: -1      default: []",
                    "$ref": "#/definitions/null.Int"
                }
            }
        },
...
        "null.Int": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "null.String": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}